Website Using the Django Web Framework
In the first project, I created an end to end website with contact form and in the second project, I created the website using an authentication system in which the user can log in and logout. This website will also handle how to authenticate the user or how to add a user. For this project, IDE used: Virtual Studio Code.